Rykket tabel funktionalitet ind i ny komponent
